test: move utils tests from test_views to test_utils
Extract TestLaunchUtils into tests/test_utils.py to match the src
structure of social_django/utils.py and clean up unused imports.
